+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed a problem which occurred on certain machined when using Sampler in
   Windows PowerShell. Fixes [#350](https://github.com/gaelcolas/Sampler/issues/350)
+- The module manifest is now correctly updated with release notes from the
+  changelog. Fixes [#358](https://github.com/gaelcolas/Sampler/issues/358)
 
 ## [0.112.3] - 2022-03-31